+++ package.json
@@ -61,6 +61,7 @@
     "postcss": "^8.5.3",
     "react-aria-components": "^1.6.0",
     "react-hook-form": "^7.54.2",
+    "recharts": "^2.15.1",
     "tailwind-merge": "^3.0.2",
     "tailwindcss": "^4.0.9",
     "tailwindcss-animate": "^1.0.7",
