+++ src/index.tsx
@@ -1,6 +1,4 @@
 import './styles/global.css';
 
-// @TODO: Only for testing purposes. It should be removed when CostCalculator is ready
-export { SelectField } from '@/src/components/form/fields/SelectField';
 export { RemoteFlows } from '@/src/RemoteFlowsProvider';
 export { ThemeProvider } from '@/src/theme';
