+++ src/lib/utils.ts
@@ -6,7 +6,6 @@ export function cn(...inputs: ClassValue[]) {
 }
 
 export function formatCurrency(amount: number, symbol = '€'): string {
-  // Divide by 100 as the amounts in the data appear to be in cents
   const value = amount / 100;
 
   return `${symbol}${value.toLocaleString('en-US', {
